Rename dual clustering file (WIP)
